refine cdc integration tests
